+++ paper/src/test/java/com/github/idankoblik/jukebox/AbstractPaperSongTest.java
@@ -13,10 +13,10 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public abstract class AbstractSpigotSongTest extends AbstractSpigotTest {
+public abstract class AbstractPaperSongTest extends AbstractPaperTest {
 
     protected static final String DEFAULT_NOTE = "note.drum";
-    protected AbstractSong song;
+    protected PaperSong song;
     protected AbstractSong locationSong;
     protected NBSSong nbsSong;
     protected PlayerMock player;
@@ -60,15 +60,15 @@ public abstract class AbstractSpigotSongTest extends AbstractSpigotTest {
         super.tearDown();
     }
 
-    protected abstract void initializeSongs();
+    protected abstract void initializeSongs() throws Exception;
 
     protected void testPlayNote() {
         // Known instrument
-        this.song.playSound((byte) 1, 1, 10);
+        this.song.playNote((byte) 1, 1);
         assertEquals(1, this.player.getHeardSounds().size());
 
         // Unknown instrument
-        this.song.playSound((byte) 69, 1, 10);
+        this.song.playNote((byte) 69, 1);
         assertEquals(2, this.player.getHeardSounds().size());
     }
 
@@ -76,21 +76,15 @@ public abstract class AbstractSpigotSongTest extends AbstractSpigotTest {
         SongState state = this.nbsSong.getState();
         assertEquals(state, SongState.IDLE);
 
-        this.song.playSong(100.f);
+        this.song.playSong();
         assertEquals(this.nbsSong.getState(), SongState.PLAYING);
         assertTrue(stageChangeListener.isWorking());
         assertEquals(stageChangeListener.getCurrent(), SongState.PLAYING);
         assertEquals(stageChangeListener.getPrevious(), SongState.IDLE);
 
-        assertTrue(startListener.isWorking());
-        try {
-            Thread.sleep(1000);
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
+        server.getScheduler().performTicks(100L);
 
         assertEquals(this.nbsSong.getState(), SongState.ENDED);
-        assertTrue(endListener.isWorking());
         assertFalse(endListener.isForcing());
 
         assertTrue(stageChangeListener.isWorking());
@@ -100,37 +94,33 @@ public abstract class AbstractSpigotSongTest extends AbstractSpigotTest {
 
     protected void testLocationPlayNote() {
         // Known instrument
-        this.locationSong.playSound((byte) 1, 1, 10);
+        this.locationSong.playNote((byte) 1, 1);
         assertEquals(0, this.player.getHeardSounds().size());
 
         // Unknown instrument
-        this.locationSong.playSound((byte) 69, 1, 10);
+        this.locationSong.playNote((byte) 69, 1);
         assertEquals(0, this.player.getHeardSounds().size());
 
         this.player.teleport(new Location(world, 5, 5,5));
 
         // Known instrument
-        this.locationSong.playSound((byte) 1, 1, 10);
+        this.locationSong.playNote((byte) 1, 1);
         assertEquals(1, this.player.getHeardSounds().size());
 
         // Unknown instrument
-        this.locationSong.playSound((byte) 69, 1, 10);
+        this.locationSong.playNote((byte) 69, 1);
         assertEquals(2, this.player.getHeardSounds().size());
     }
 
     protected void testPlaySong() {
-        Song song = (Song) this.song;
-        assertFalse(song.isLoop());
+        PaperSong song = (PaperSong) this.song;
+        assertFalse(song.toLoop());
         song.setLoop(true);
-        assertTrue(song.isLoop());
-        song.playSong(100.f);
-        try {
-            Thread.sleep(1000);
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
+        assertTrue(song.toLoop());
+        song.playSong();
+        server.getScheduler().performTicks(4);
         song.stopSong();
-        assertEquals(19, this.player.getHeardSounds().size());
+        assertEquals(3, this.player.getHeardSounds().size());
         assertEquals(this.nbsSong.getState(), SongState.ENDED);
         assertTrue(endListener.isWorking());
         assertTrue(endListener.isForcing());